+++ datanode/entities/party_vesting_balances.go
@@ -4,6 +4,7 @@ import (
 	"time"
 
 	"code.vegaprotocol.io/vega/libs/num"
+	eventspb "code.vegaprotocol.io/vega/protos/vega/events/v1"
 )
 
 type (
@@ -24,3 +25,44 @@ type (
 		VegaTime time.Time
 	}
 )
+
+func PartyVestingBalanceFromProto(
+	partyID string,
+	atEpoch uint64,
+	pvb *eventspb.PartyVestingBalance,
+	t time.Time,
+) (*PartyVestingBalance, error) {
+	balance, err := num.DecimalFromString(pvb.Balance)
+	if err != nil {
+		return nil, err
+	}
+
+	return &PartyVestingBalance{
+		PartyID:  PartyID(partyID),
+		AssetID:  AssetID(pvb.Asset),
+		AtEpoch:  atEpoch,
+		Balance:  balance,
+		VegaTime: t,
+	}, nil
+}
+
+func PartyLockedBalanceFromProto(
+	partyID string,
+	atEpoch uint64,
+	pvb *eventspb.PartyLockedBalance,
+	t time.Time,
+) (*PartyLockedBalance, error) {
+	balance, err := num.DecimalFromString(pvb.Balance)
+	if err != nil {
+		return nil, err
+	}
+
+	return &PartyLockedBalance{
+		PartyID:    PartyID(partyID),
+		AssetID:    AssetID(pvb.Asset),
+		AtEpoch:    atEpoch,
+		UntilEpoch: pvb.UntilEpoch,
+		Balance:    balance,
+		VegaTime:   t,
+	}, nil
+}